🧹clean: remove all unwanted resources
