+++ src/components/ContentLayout.tsx
@@ -3,11 +3,7 @@ import { twMerge } from "tailwind-merge";
 
 export const ContentLayout = forwardRef<HTMLDivElement, ComponentProps<"div">>(
   ({ children, className, ...props }, ref) => (
-    <div
-      ref={ref}
-      className={twMerge("flex flex-row h-[calc(100vh-28px)] mt-7", className)}
-      {...props}
-    >
+    <div ref={ref} className={twMerge("flex flex-row h-screen", className)} {...props}>
       {children}
     </div>
   ),
