+++ src/components/body/Available.tsx
@@ -1,27 +1,21 @@
-import { loadNodeVersionAtom, selectedToolAtom } from "@/store";
-import { useAtomValue, useSetAtom } from "jotai";
+import { loadNodeVersionAtom } from "@/store";
+import { useSetAtom } from "jotai";
 import { getNodeVersionList, VersionType } from "@/libs";
 import { List } from "./list";
 import { SymbolIcon } from "@radix-ui/react-icons";
 import { Button } from "@/components/Button";
-import { useEffect } from "react";
 
 export const Available = () => {
   const loadNodeVersion = useSetAtom(loadNodeVersionAtom);
-  const selectedTool = useAtomValue(selectedToolAtom);
   const all = getNodeVersionList(VersionType.DEFAULT);
 
   const refreshAvailableVersion = async (isRefresh: boolean) => {
     await loadNodeVersion(isRefresh);
   };
 
-  useEffect(() => {
-    refreshAvailableVersion(false);
-  }, [selectedTool]);
-
   return (
     <div className="flex flex-col items-center justify-center mt-5">
-      <title>{selectedTool}</title>
+      <title>Node</title>
       <div className="p-2 font-black rounded-md text-card-foreground bg-card shadow-md flex items-center justify-center mb-4 text-xl gap-x-2 lg:text-4xl">
         Available For Installation
         <Button
@@ -38,7 +32,7 @@ export const Available = () => {
             className="p-4 text-orange-700 bg-orange-100 border-l-4 border-orange-500"
             role="alert"
           >
-            <p className="font-bold">No List available for {selectedTool} to install</p>
+            <p className="font-bold">No List available for Node to install</p>
             <p>Try refreshing the list.</p>
           </div>
         ) : (
