+++ src/components/body/index.ts
@@ -1,2 +1,5 @@
-export * from "./node";
+export * from "./list";
+
+export * from "./Installed";
+export * from "./Available";
 export * from "./ContentBody";
