--- src/components/body/node/Node.tsx
@@ -1,11 +1,0 @@
-import { Available } from "./Available";
-import { Installed } from "./Installed";
-
-export const Node = () => {
-  return (
-    <>
-      <Installed />
-      <Available />
-    </>
-  );
-};
