--- src/components/body/node/index.ts
@@ -1,5 +1,0 @@
-export * from "./list";
-
-export * from "./Installed";
-export * from "./Available";
-export * from "./Node";
