+++ src/components/index.ts
@@ -2,4 +2,4 @@ export * from "./Button";
 export * from "./ContentLayout";
 
 export * from "./body";
-export * from "./titlebar";
+export * from "./controls";
