--- src/components/titlebar/Center.tsx
@@ -1,5 +1,0 @@
-export const Center = () => {
-  return (
-    <div className="font-black cursor-default text-text">JavaScript Version Manager</div>
-  );
-};
