+++ src/libs/constants/index.ts
@@ -1,4 +1,3 @@
-export * from "./tool-list";
 export * from "./config";
 export * from "./url";
 export * from "./font-list";
