+++ src/libs/enums/index.ts
@@ -1,3 +1,3 @@
-export * from "./tool";
+export * from "./version-type";
 export * from "./file-sys-res";
 export * from "./toaster-type";
