+++ src/libs/enums/version-type.ts
@@ -1,9 +1,3 @@
-export enum ToolType {
-  NODE = "Node",
-  BUN = "Bun",
-  DENO = "Deno",
-}
-
 export enum VersionType {
   LTS = "LTS",
   CURRENT = "Current",
