+++ src/libs/models/index.ts
@@ -1,4 +1,3 @@
-export * from "./tool-list";
 export * from "./config";
 export * from "./node-version";
 export * from "./toaster";
