--- src/libs/models/tool-list.ts
@@ -1,8 +1,0 @@
-import { ToolType } from "../enums";
-
-export interface ToolList {
-  name: string;
-  icon: string;
-  type: ToolType;
-  style?: string;
-}
