+++ src/store/app.ts
@@ -1,18 +1,9 @@
-import { ToolType } from "@/libs";
 import { atom } from "jotai";
 
-export const selectedToolAtom = atom<ToolType | null>(null);
 export const searchTermAtom = atom<string>("");
 export const downloadingVersionAtom = atom<string>("");
 export const toastIdAtom = atom<number>(0);
 export const downloadingProgressAtom = atom<number>(0);
-
-export const updateSelectedToolAtom = atom(
-  null,
-  async (_get, set, type: ToolType | null) => {
-    set(selectedToolAtom, type);
-  },
-);
 
 export const updateSearchTermAtom = atom(null, async (_get, set, searchTerm: string) => {
   set(searchTermAtom, searchTerm);
